move clone logic out of LocalRepository into the command::clone module
