+++ src/lib/src/command/clone.rs
@@ -5,13 +5,17 @@
 
 use std::path::Path;
 
-use crate::constants::DEFAULT_BRANCH_NAME;
+use crate::config::RemoteConfig;
+use crate::constants::{DEFAULT_BRANCH_NAME, DEFAULT_REMOTE_NAME, REPO_CONFIG_FILENAME};
+use crate::core::index::EntryIndexer;
 use crate::error::OxenError;
-use crate::model::LocalRepository;
-use crate::opts::CloneOpts;
+use crate::model::{LocalRepository, Remote, RemoteBranch, RemoteRepository};
+use crate::opts::{CloneOpts, PullOpts};
+use crate::util::progress_bar::{oxen_progress_bar, ProgressBarType};
+use crate::{api, util};
 
 pub async fn clone(opts: &CloneOpts) -> Result<LocalRepository, OxenError> {
-    match LocalRepository::clone_remote(opts).await {
+    match clone_remote(opts).await {
         Ok(Some(repo)) => Ok(repo),
         Ok(None) => Err(OxenError::remote_repo_not_found(&opts.url)),
         Err(err) => Err(err),
@@ -61,6 +65,136 @@ async fn _clone(
     clone(&opts).await
 }
 
+async fn clone_remote(opts: &CloneOpts) -> Result<Option<LocalRepository>, OxenError> {
+    log::debug!(
+        "clone_remote {} -> {:?} -> shallow? {} -> all? {}",
+        opts.url,
+        opts.dst,
+        opts.shallow,
+        opts.all
+    );
+
+    let remote = Remote {
+        name: String::from(DEFAULT_REMOTE_NAME),
+        url: opts.url.to_owned(),
+    };
+    let remote_repo = api::remote::repositories::get_by_remote(&remote)
+        .await?
+        .ok_or_else(|| OxenError::remote_repo_not_found(&opts.url))?;
+    let repo = clone_repo(remote_repo, opts).await?;
+    Ok(Some(repo))
+}
+
+async fn clone_repo(
+    remote_repo: RemoteRepository,
+    opts: &CloneOpts,
+) -> Result<LocalRepository, OxenError> {
+    api::remote::repositories::pre_clone(&remote_repo).await?;
+
+    // if directory already exists -> return Err
+    let repo_path = &opts.dst;
+    if repo_path.exists() {
+        let err = format!("Directory already exists: {}", remote_repo.name);
+        return Err(OxenError::basic_str(err));
+    }
+
+    // if directory does not exist, create it
+    std::fs::create_dir_all(repo_path)?;
+
+    // if create successful, create .oxen directory
+    let oxen_hidden_path = util::fs::oxen_hidden_dir(repo_path);
+    std::fs::create_dir(&oxen_hidden_path)?;
+
+    // save LocalRepository in .oxen directory
+    let repo_config_file = oxen_hidden_path.join(Path::new(REPO_CONFIG_FILENAME));
+    let mut local_repo = LocalRepository::from_remote(remote_repo.clone(), repo_path)?;
+    repo_path.clone_into(&mut local_repo.path);
+    local_repo.set_remote(DEFAULT_REMOTE_NAME, &remote_repo.remote.url);
+
+    // Save remote config in .oxen/config.toml
+    let remote_cfg = RemoteConfig {
+        remote_name: Some(DEFAULT_REMOTE_NAME.to_string()),
+        remotes: vec![remote_repo.remote.clone()],
+    };
+
+    let toml = toml::to_string(&remote_cfg)?;
+    util::fs::write_to_path(&repo_config_file, &toml)?;
+
+    // Pull all commit objects, but not entries
+    let rb = RemoteBranch::from_branch(&opts.branch);
+    let indexer = EntryIndexer::new(&local_repo)?;
+    maybe_pull_entries(&local_repo, &remote_repo, &indexer, &rb, opts).await?;
+
+    if opts.all {
+        log::debug!("pulling all entries");
+        let remote_branches = api::remote::branches::list(&remote_repo).await?;
+        if remote_branches.len() > 1 {
+            println!(
+                "🐂 Pre-fetching {} additional remote branches...",
+                remote_branches.len() - 1
+            );
+        }
+
+        let n_other_branches: u64 = if remote_branches.len() > 1 {
+            (remote_branches.len() - 1) as u64
+        } else {
+            0
+        };
+
+        let bar = oxen_progress_bar(n_other_branches as u64, ProgressBarType::Counter);
+
+        for branch in remote_branches {
+            // We've already pulled the target branch in full
+            if branch.name == rb.branch {
+                continue;
+            }
+
+            let remote_branch = RemoteBranch::from_branch(&branch.name);
+            indexer
+                .pull_most_recent_commit_object(&remote_repo, &remote_branch, false)
+                .await?;
+            bar.inc(1);
+        }
+        bar.finish_and_clear();
+    }
+
+    println!(
+        "\n🎉 cloned {} to {}/\n",
+        remote_repo.remote.url, remote_repo.name
+    );
+    api::remote::repositories::post_clone(&remote_repo).await?;
+
+    Ok(local_repo)
+}
+
+async fn maybe_pull_entries(
+    local_repo: &LocalRepository,
+    remote_repo: &RemoteRepository,
+    indexer: &EntryIndexer,
+    rb: &RemoteBranch,
+    opts: &CloneOpts,
+) -> Result<(), OxenError> {
+    // Shallow means we will not pull the actual data until a user tells us to
+    if opts.shallow {
+        indexer
+            .pull_most_recent_commit_object(remote_repo, rb, true)
+            .await?;
+        local_repo.write_is_shallow(true)?;
+    } else {
+        // Pull all entries
+        indexer
+            .pull(
+                rb,
+                PullOpts {
+                    should_pull_all: opts.all,
+                    should_update_head: true,
+                },
+            )
+            .await?;
+    }
+    Ok(())
+}
+
 #[cfg(test)]
 mod tests {
     use crate::api;
@@ -72,6 +206,72 @@ mod tests {
     use crate::model::RepoNew;
     use crate::test;
     use crate::util;
+
+    use super::*;
+
+    #[tokio::test]
+    async fn test_clone_remote() -> Result<(), OxenError> {
+        test::run_empty_local_repo_test_async(|local_repo| async move {
+            // Create remote repo
+            let remote_repo = test::create_remote_repo(&local_repo).await?;
+
+            log::debug!("created the remote repo");
+
+            test::run_empty_dir_test_async(|dir| async move {
+                let opts = CloneOpts::new(remote_repo.remote.url.to_owned(), dir.join("new_repo"));
+
+                log::debug!("about to clone the remote");
+                let local_repo = clone_remote(&opts).await?.unwrap();
+                log::debug!("succeeded");
+                let cfg_fname = ".oxen/config.toml".to_string();
+                let config_path = local_repo.path.join(&cfg_fname);
+                assert!(config_path.exists());
+
+                let repository = LocalRepository::from_dir(&local_repo.path);
+                assert!(repository.is_ok());
+
+                let repository = repository.unwrap();
+                let status = command::status(&repository)?;
+                assert!(status.is_clean());
+
+                // Cleanup
+                api::remote::repositories::delete(&remote_repo).await?;
+
+                Ok(dir)
+            })
+            .await
+        })
+        .await
+    }
+
+    #[tokio::test]
+    async fn test_move_local_repo_path_valid() -> Result<(), OxenError> {
+        test::run_empty_local_repo_test_async(|local_repo| async move {
+            // Create remote repo
+            let remote_repo = test::create_remote_repo(&local_repo).await?;
+
+            test::run_empty_dir_test_async(|dir| async move {
+                let opts = CloneOpts::new(remote_repo.remote.url.to_owned(), dir.join("new_repo"));
+                let local_repo = clone_remote(&opts).await?.unwrap();
+
+                api::remote::repositories::delete(&remote_repo).await?;
+
+                command::status(&local_repo)?;
+
+                let new_path = dir.join("new_path");
+
+                util::fs::rename(&local_repo.path, &new_path)?;
+
+                let new_repo = LocalRepository::from_dir(&new_path)?;
+                command::status(&new_repo)?;
+                assert_eq!(new_repo.path, new_path);
+
+                Ok(dir)
+            })
+            .await
+        })
+        .await
+    }
 
     // Test for clone --all that checks to make sure we have all commits, all deleted files, etc
     #[tokio::test]
